Fixed a few trivial compiler errors
